+++ example/config/dev.js
