+++ example/src/ClientConfig.js
