+++ example/src/Query.js
