+++ example/src/Query.universal.js
