+++ example/src/admin/noot_static.html.js
