+++ example/src/universalComponent.js
