+++ injects/browser.js
