+++ injects/server.js
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import * as fs from 'fs'
 
-export { fs, cx, React }
+export { cx, React, fs }
 
 function cx(...args) {
   return args.filter(Boolean).join(' ')
