+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@kaliber/config": "^0.0.8",
     "esbuild": "^0.21.4",
+    "gettext-parser": "^6.0.0",
     "hoist-non-react-statics": "^3.3.2",
     "import-fresh": "^3.3.0"
   },
